Change return type of `parse_k_points` to 5 possibilities
Why this change?
In the previous change, if we use
`a, b = QuantumESPRESSOParsers.OutputParsers.PWscf.parse_k_points(str)`,
if no crystal coordinates present, the return type will be a `Matrix`.
Thus, due to Julia's destruction behavior, `a`, `b` will be the
first 2 elements of the matrix. If a user writes the code like above,
he/she will probably get different return type when parsing
different text. This is not expected.
This change make it always return a `nothing` (no regex found) or a `Tuple`,
where the tuple have 4 possibilities:
`(Nothing, Nothing)`, `(Matrix, Nothing)`, `(Nothing, Matrix)`, `(Matrix, Matrix)`.
So `a` and `b` will always be `nothing` or a `Matrix`.